refactor(metering): move tests inline with MeteringTestContext
- Add MeteringTestContext for unit tests that don't need a full node
- Move meter.rs tests inline using MeteringTestContext (sync tests)
- Move block.rs tests inline using MeteringTestContext (sync tests)
- Move rpc.rs tests inline using TestHarness with MeteringExtension
- Delete tests/ directory (all tests now inline)
- Update lib.rs to expose test_utils with #[cfg(any(test, feature = "test-utils"))]
