+++ crates/client/metering/src/lib.rs
@@ -20,3 +20,6 @@ pub use traits::MeteringApiServer;
 
 mod types;
 pub use types::{MeterBlockResponse, MeterBlockTransactions};
+
+#[cfg(any(test, feature = "test-utils"))]
+pub mod test_utils;
